requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
flask_marshmallow==0.14.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
SQLAlchemy==1.4.31
PyJWT==2.3.0
mailjet_rest==1.3.4
requests==2.27.1